Admin Map
- City Model
- Point Model
- Map Controller
- migrations fixes
- CitiesTableSeeder
- Index View loop

+++ database/migrations/2017_04_05_152302_create_table_cities.php
@@ -16,8 +16,8 @@ class CreateTableCities extends Migration
         Schema::create('cities', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name');
-            $table->integer('center-lat');
-            $table->integer('center-long');
+            $table->float('center-lat', 9, 6);
+            $table->float('center-long', 9, 6);
             $table->timestamps();
         });
     }
